# docs/swagger.yaml
definitions:
...
  models.RequestEvent:
    properties:
      approval_date:
        type: string
      approver_id:
        type: integer
      isbn:
        type: string
      issue_status:
        description: e.g., "Issued and Approved"
        type: string
      reader_id:
        type: integer
      req_id:
        type: integer
      request_date:
        type: string
      request_type:
        description: e.g., "Borrow", "Return"
        type: string
    required:
    - isbn
    type: object
...
info:
  contact: {}
paths:
  /approve-issue/{id}:
    put:
      consumes:
      - application/json
      description: Approve a book request and issue the book to the user
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book issued and approved successfully
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Request not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: No available copies for this book
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update book inventory or issue book
          schema:
            additionalProperties: true
            type: object
      summary: Approve and issue a book request
      tags:
      - requests
...
  /request:
    post:
      consumes:
      - application/json
      description: Create a new request for a book
      parameters:
      - description: Request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestEvent'
      produces:
      - application/json
      responses:
        "201":
          description: Request created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Cannot Bind JSON Data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Book not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Request already exists or Book is not available
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Cannot create request
          schema:
            additionalProperties: true
            type: object
      summary: Create a new request
      tags:
      - requests
  /requests:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all requests
      produces:
      - application/json
      responses:
        "200":
          description: All requests retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch requests
          schema:
            additionalProperties: true
            type: object
      summary: Get all requests
      tags:
      - requests
...
swagger: "2.0"